+++ extensions/downloads-manager/src/utils.tsx
@@ -1,5 +1,7 @@
 import { Action, ActionPanel, confirmAlert, Detail, getPreferenceValues, showToast, Toast } from "@raycast/api";
-import { accessSync, rmSync, constants, readdirSync, statSync } from "fs";
+import { showFailureToast } from "@raycast/utils";
+import { accessSync, constants, readdirSync, statSync } from "fs";
+import { rm } from "fs/promises";
 import { join } from "path";
 import { ComponentType } from "react";
 import untildify from "untildify";
@@ -71,7 +73,7 @@ export function hasAccessToDownloadsFolder() {
 export async function deleteFileOrFolder(filePath: string) {
   const shouldDelete = await confirmAlert({
     title: "Delete Item?",
-    message: `Are you sure you want to delete:\n${filePath}?`,
+    message: `Are you sure you want to permanently delete:\n${filePath}?`,
     primaryAction: {
       title: "Delete",
     },
@@ -83,11 +85,11 @@ export async function deleteFileOrFolder(filePath: string) {
   }
 
   try {
-    rmSync(filePath, { recursive: true, force: true });
+    rm(filePath, { recursive: true, force: true });
     await showToast({ style: Toast.Style.Success, title: "Item Deleted" });
   } catch (error) {
     if (error instanceof Error) {
-      await showToast({ style: Toast.Style.Failure, title: "Deletion Failed", message: error.message });
+      await showFailureToast(error, { title: "Deletion Failed" });
     }
   }
 }